+++ src/js/gutenberg/blocks/static-bibliography/save.tsx
@@ -1,5 +1,5 @@
+import { RichText } from '@wordpress/block-editor';
 import { BlockSaveProps } from '@wordpress/blocks';
-import { RichText } from '@wordpress/editor';
 import { Component } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';